Cleanup for memory efficient fusion
