+++ benchmarks/transformer_fusion_patterns/benchmark_helper.py
