+++ functorch/compile/__init__.py
@@ -1,5 +1,4 @@
 from .._src.operator_authoring import pointwise_operator
-from .._src.memory_efficient_op_authoring import memory_efficient_pointwise_fusion, torchscript_nvfuser_compile
 from .._src.python_key import pythonkey_decompose, pythonkey_meta
 from .._src.decompositions import register_decomposition, decomposition_table
 from .._src.fx_minifier import minimizer, check_nvfuser_subprocess
@@ -15,4 +14,4 @@ from .._src.aot_autograd import (
     clear_compile_cache,
     draw_graph,
 )
-from .._src.compilers import ts_compile, tvm_compile, draw_graph_compile, nop, nnc_jit
+from .._src.compilers import ts_compile, tvm_compile, draw_graph_compile, nop, nnc_jit, memory_efficient_fusion
